update2BoltFastAPI/project/backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.6.1
pdfminer.six==20231228
deepgram-sdk==2.12.0
moviepy==1.0.3
ffmpeg-python==0.2.0
assemblyai==0.17.0
together==0.2.9
requests==2.31.0